requirements.txt
Flask==3.0.3
python-docx==1.1.2
gunicorn==22.0.0
itsdangerous==2.2.0
jinja2==3.1.4
Werkzeug==3.0.3